feat: 프로젝트 페이지 UIUX 구현
