--- src/app/(home)/template.tsx
@@ -1,6 +1,0 @@
-import { ReactNode } from "react";
-import TransitionLayout from "../TransitionLayout";
-
-export default function AboutTemplate({ children }: { children: ReactNode }) {
-  return <TransitionLayout>{children}</TransitionLayout>;
-}
