+++ src/app/TransitionLayout.tsx
@@ -20,7 +20,6 @@ export default function TransitionLayout({
         animate={{ opacity: 1, y: 0 }}
         exit={{ opacity: 0, y: -10 }}
         transition={{ duration: 0.3, ease: "easeOut" }}
-        className={clsx("flex-1 overflow-x-hidden")}
       >
         {children}
       </motion.div>
