+++ src/app/projects/types.ts
@@ -1,0 +1,4 @@
+export type ProjectContentsProps = {
+  $frontend?: boolean;
+  $publishing?: boolean;
+};
