--- src/data/frontend/types.ts
@@ -1,11 +1,0 @@
-export type ProntendData = {
-  id: number;
-  logoDefaultImg: string;
-  logoActiveImg: string;
-  thumbnailImg: string;
-  keyword: string[];
-  title: string;
-  description: string;
-  exploreLink:string;
-  diveInLink:string;
-};
