+++ src/data/home/frontend/types.ts
@@ -1,0 +1,11 @@
+export type ProntendData = {
+  id: number;
+  logoDefaultImg: string;
+  logoActiveImg: string;
+  thumbnailImg: string;
+  keyword: string[];
+  title: string;
+  description: string;
+  exploreLink:string;
+  diveInLink:string;
+};
