+++ src/data/home/techBlog/types.ts
@@ -1,0 +1,8 @@
+export type TechPost = {
+  id: number;
+  title: string;
+  excerpt: string;
+  category: string[];
+  date: string;
+  link: string;
+};
