--- src/data/publishing/index.ts
@@ -1,81 +1,0 @@
-import { PublishingData } from "./types";
-
-export const publishingProject: PublishingData[] = [
-  {
-    id: 1,
-    logoImg: "/images/home/publishingSection/logo/portfolio.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/portfolio.png",
-    keyword: ["html", "css", "jquery", "portfolio", "publishing", "responsive"],
-    title: "Publisher Portfolio",
-    description:
-      "HTML, CSS, jQuery 기반 개인 포트폴리오 / 디자인·퍼블리싱 작업물 수록",
-    exploreLink: "",
-    diveInLink: "https://publisher-nerimy.netlify.app/",
-  },
-  {
-    id: 2,
-    logoImg: "/images/home/publishingSection/logo/becelo.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/becelo.png",
-    keyword: [
-      "company_collaboration",
-      "php",
-      "javascript",
-      "jquery",
-      "publishing",
-      "responsive",
-      "web&app",
-    ],
-    title: "Becelo",
-    description:
-      "PHP 기반 레거시 프로젝트의 판매자·일반 사용자 앱, 관리자·랜딩 페이지 전체 퍼블리싱 리뉴얼",
-    exploreLink: "",
-    diveInLink: "https://becelo.com/bbs/login.php",
-  },
-  {
-    id: 3,
-    logoImg: "/images/home/publishingSection/logo/fansign.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/fansign.png",
-    keyword: ["company_collaboration", "flutter", "dart", "publishing"],
-    title: "FANSIGN",
-    description:
-      "Flutter(Dart) 기반 앱 개발 참여 / 앱 화면 구성과 코드 설명을 정리한 HTML 가이드 페이지 제작",
-    exploreLink: "",
-    diveInLink: "https://publisher-nerimy.netlify.app/port/fansign_app/",
-  },
-  {
-    id: 4,
-    logoImg: "/images/home/publishingSection/logo/travelyMap.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/travelyMap.png",
-    keyword: ["company_collaboration", "react", "publishing", "web&app"],
-    title: "travelymap",
-    description:
-      "React 기반 웹앱 프로젝트에 퍼블리셔로 참여 / 퍼블리싱 코드 작성 및 GitHub으로 코드 관리·유지보수 진행",
-    exploreLink: "",
-    diveInLink:
-      "https://publisher-nerimy.netlify.app/port/travelymap_app/guidemagazine/guide_home",
-  },
-  {
-    id: 5,
-    logoImg: "/images/home/publishingSection/logo/homtStudio.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/homtstudio.png",
-    keyword: ["company_collaboration", "publishing", "web&app"],
-    title: "HOMT STUDIO",
-    description:
-      "HTML, CSS 기반 웹앱 퍼블리싱 / 개발 연동용 퍼블리싱 코드 작업 및 전달",
-    exploreLink: "",
-    diveInLink:
-      "https://publisher-nerimy.netlify.app/port/homtstudio_app/login/login",
-  },
-  {
-    id: 6,
-    logoImg: "/images/home/publishingSection/logo/ketock.svg",
-    thumbnailImg: "/images/home/publishingSection/thumbnail/ketock.png",
-    keyword: ["company_collaboration", "publishing", "web&app"],
-    title: "KETOCK",
-    description:
-      "HTML, CSS 기반 웹앱 퍼블리싱 / 개발 연동용 퍼블리싱 코드 작업 및 전달",
-    exploreLink: "",
-    diveInLink:
-      "https://publisher-nerimy.netlify.app/port/ketock_app/login/login1",
-  },
-];
