--- src/data/publishing/types.ts
@@ -1,10 +1,0 @@
-export type PublishingData = {
-  id: number;
-  logoImg: string;
-  thumbnailImg: string;
-  keyword: string[];
-  title: string;
-  description: string;
-  exploreLink: string;
-  diveInLink: string;
-};
