--- src/data/techBlog/types.ts
@@ -1,8 +1,0 @@
-export type TechPost = {
-  id: number;
-  title: string;
-  excerpt: string;
-  category: string[];
-  date: string;
-  link: string;
-};
