+++ src/styles/globals.css
@@ -325,3 +325,23 @@ body {
     background-position: 0 -65px;
   }
 }
+
+.slick-dots {
+  position: relative;
+  bottom: -15px;
+}
+
+.slick-dots li button:before {
+  font-size: 15px;
+  content: "■";
+}
+
+.slick-dots li.slick-active button:before {
+  opacity: 1;
+  color: white;
+}
+
+.slick-dots li button:before {
+  opacity: 0.4;
+  color: var(--white);
+}
